requirements.txt
absl-py==0.8.1
astor==0.8.1
backcall==0.1.0
biopython==1.70
boto3==1.10.19
botocore==1.13.19
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
-e git+https://github.com/flatironinstitute/DeepFRI.git@0eab9bbeb24981cae22c5261803c3cfaf5168f65#egg=DeepFRIer
docutils==0.15.2
entrypoints==0.3
fairseq==0.9.0
fastBPE==0.1.0
flake8==3.7.9
gast==0.3.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
jmespath==0.9.4
joblib==0.14.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.3
mccabe==0.6.1
mmdnn==0.2.5
networkx==2.1
nose==1.3.7
numpy==1.18.1
oauthlib==3.1.0
onnx==1.6.0
onnx2keras==0.0.4
pandas==0.25.3
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
-e git+https://github.com/mortonjt/p2p.git@288b2afdb4f0d9394b7155f3ffd02f91f87e56fa#egg=poplar
portalocker==1.5.2
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.0
pytorch2keras==0.2.3
pytz==2019.3
PyYAML==5.3
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
sacrebleu==1.4.2
sacremoses==0.0.35
scikit-learn==0.19.0
scipy==1.4.1
seaborn==0.10.0
sentencepiece==0.1.83
six==1.13.0
statsmodels==0.11.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.38.0
traitlets==4.3.3
transformers==2.1.1
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
